ReverseShell file extensions
When opening a reverse shell connection from a web site, to understand the behavior of the site, this list can be used. Web sites can be seduced by using different types of file exceptions. For example, a website can reject .php file, but it can accept .php5 files. This list can be developed, and it can be used in BurpSuite Intruder.